Instead of "play" button, added a button "transfer/upload to device"
