+++ app/src/main/res/layout-land/activity_main.xml
@@ -46,17 +46,4 @@
         app:labelVisibilityMode="selected"
         app:menu="@menu/navigation" />
 
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fab_main"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/fab_margin"
-        android:layout_marginTop="@dimen/fab_margin"
-        android:layout_marginEnd="@dimen/fab_margin"
-        android:layout_marginBottom="@dimen/bottom_bar_margin"
-        android:tint="@android:color/white"
-        app:layout_anchor="@id/mainLayout"
-        app:layout_anchorGravity="bottom|right"
-        app:srcCompat="@drawable/ic_send" />
-
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
